# internal/tds-coaching-change-service/handlers/kubernetes.go
// Package handlers TDS Coaching Change Service API
//
// Documentation of TDS Coaching Change Service API.
//
//     Schemes: http
//     BasePath: /
//     Version: 1.0.0
//     Host: localhost
//
//     Consumes:
//     - application/json
//
//     Produces:
//     - application/json
//
// swagger:meta

package handlers

import (
	"encoding/json"
	"github.com/ocrosby/soccer/internal/tds-coaching-change-service/models"
	"log"
	"net/http"
)

// KubernetesHandler is a handler for the kubernetes endpoints
type KubernetesHandler struct {
}

// NewKubernetesHandler creates a new KubernetesHandler
func NewKubernetesHandler() *KubernetesHandler {
	return &KubernetesHandler{}
}

func isHealthy() bool {
	// Your health check logic here
	return true
}

func isReady() bool {
	// Your readiness check logic here
	return true
}

func isStartupReady() bool {
	// Your startup check logic here
	return true
}

// swagger:route GET /health Kubernetes healthCheck
// HealthCheckHandler checks if the service is healthy.
// responses:
//
//	200: HealthResponse
//	503: HealthResponse
func (h *KubernetesHandler) HealthCheckHandler(w http.ResponseWriter, r *http.Request) {
	// Your logic to check if the service is healthy
	// If healthy:
	healthy := isHealthy()

	w.Header().Set("Content-Type", "application/json")

	if healthy {
		response := models.HealthResponse{Status: "healthy"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			log.Printf("Error encoding response: %v", err)
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		}
	} else {
		response := models.HealthResponse{Status: "unhealthy"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			log.Printf("Error encoding response: %v", err)
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		} else {
			http.Error(w, "Service Unavailable", http.StatusServiceUnavailable)
		}
	}
}

// swagger:route GET /readiness Kubernetes readinessCheck
// ReadinessCheckHandler checks if the service is ready to serve traffic.
// responses:
//
//	200: ReadinessResponse
//	503: ReadinessResponse
func (h *KubernetesHandler) ReadinessCheckHandler(w http.ResponseWriter, r *http.Request) {
	// Example readiness check (replace with actual checks)
	ready := isReady() // Assume readiness checks are implemented and result in this boolean

	w.Header().Set("Content-Type", "application/json")

	if ready {
		response := models.ReadinessResponse{Status: "ready"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		}
	} else {
		response := models.ReadinessResponse{Status: "not ready"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			log.Printf("Error encoding readiness response: %v", err)
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		} else {
			http.Error(w, "Service Unavailable", http.StatusServiceUnavailable)
		}
	}
}

// swagger:route GET /startup Kubernetes startupCheck
// StartupCheckHandler checks if the service has completed its startup procedures.
// responses:
//
//	200: StartupResponse
//	503: StartupResponse
func (h *KubernetesHandler) StartupCheckHandler(w http.ResponseWriter, r *http.Request) {
	// Initialize startup status to false to simulate startup checks
	startupReady := isStartupReady()

	w.Header().Set("Content-Type", "application/json")

	if startupReady {
		response := models.StartupResponse{Status: "ready"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			log.Printf("Error encoding startup response: %v", err)
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		}
	} else {
		response := models.StartupResponse{Status: "not ready"}
		if err := json.NewEncoder(w).Encode(response); err != nil {
			log.Printf("Error encoding startup response: %v", err)
			http.Error(w, "Internal server error", http.StatusInternalServerError)
			return
		} else {
			http.Error(w, "Service Unavailable", http.StatusServiceUnavailable)
		}
	}
}